+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Changelog
 Notable changes to this project.
 
+## [0.5] - development
+- complete rewrite after many tournament changes
+
 ## [0.4.1] - 2019-10-01
 - added log of tournament changes (`settings.events`)
 - use event log instead of magic numbers
 - increase test coverage to >90%
-- Fix missing columns and data fetch for rounds using correclation score (@jrdi)
+- Fix missing columns and data fetch for rounds using correlation score (@jrdi)
 
 ## [0.3.3 - 2019-07-27]
 - `reputation` report gained a `rank` option, that allows showing the